requirements.txt
Flask~=3.1.0
Werkzeug~=3.1.3
groq~=0.15.0
mediapipe==0.10.20
numpy~=1.26.4
flask-cors
python-dotenv~=1.0.1
opencv-python-headless~=4.9.0.80
